fix: analise e correcoes de erros e inconsistencias
- Frontend: tratar falhas silenciosas em TechContractInfoPage (historico de visitas) e TechWorkOrderDetailPage (chats/notas) com toast e estado vazio
- Backend: try/catch e log em controllers de delete (ServiceOps, Central, CrmFeatures, CrmLeadScoringTrait, FinanceAdvanced, HRAdvanced)
- CommissionService: log em nivel debug quando trigger de comissao e ignorado
- Migration quote_module: comentario no catch do down() para rollback idempotente
- Demais alteracoes pre-existentes no repo (migrations, modelos, rotas, frontend)

+++ backend/app/Http/Controllers/Api/V1/InventoryController.php
@@ -228,7 +228,7 @@ class InventoryController extends Controller
 
     private function authorizeTenant(Inventory $inventory)
     {
-        if ($inventory->tenant_id !== app('current_tenant_id')) {
+        if ($inventory->tenant_id !== (int) app('current_tenant_id')) {
             abort(403);
         }
     }
